move urlUtils outside components folder
